Refactor and reorganize 2 routes
